--- Procfile
@@ -1,1 +1,0 @@
-web: sh -c 'python manage.py migrate && gunicorn web/hack.wsgi'
