--- core/static/core/assets/js/main.js
@@ -1,160 +1,0 @@
-/*=============== SHOW MENU ===============*/
-const navMenu = document.getElementById('nav-menu'),
-    navToggle = document.getElementById('nav-toggle')
-navClose = document.getElementById('nav-close')
-
-/*===== MENU SHOW =====*/
-/* Validate if constant exists */
-if (navToggle) {
-    navToggle.addEventListener('click', () => {
-        navMenu.classList.add('show-menu');
-    })
-}
-
-/*===== MENU HIDDEN =====*/
-/* Validate if constant exists */
-if (navClose) {
-    navClose.addEventListener('click', () => {
-        navMenu.classList.remove('show-menu');
-    })
-}
-
-/*=============== REMOVE MENU ON LINK CLICK ===============*/
-const navLink = document.querySelectorAll('.nav__link');
-
-function linkClicked() {
-    navMenu.classList.remove('show-menu');
-}
-// Remove the Nav Menu when a link is clicked.
-navLink.forEach(link => link.addEventListener('click', linkClicked));
-
-
-/*=============== CHANGE BACKGROUND HEADER ===============*/
-function scrollHeader() {
-    const header = document.getElementById('header');
-    // When the scroll is greater than viewport height, add the scroll-header class to the header tag
-    if (this.scrollY >= 80) header.classList.add('scroll-header');
-    else header.classList.remove('scroll-header');
-}
-
-window.addEventListener('scroll', scrollHeader);
-
-
-/*=============== QUESTIONS ACCORDION ===============*/
-const accordionItems = document.querySelectorAll(".questions__item");
-
-accordionItems.forEach(item => {
-    const accordionHeader = item.querySelector(".question__header");
-
-    accordionHeader.addEventListener("click", () => {
-
-        const openItem = document.querySelector('.accordion-open');
-
-        toggleItem(item);
-
-        if (openItem && openItem !== item) {
-            toggleItem(openItem);
-        }
-    })
-})
-
-const toggleItem = (item) => {
-    const accordionContent = item.querySelector(".questions__content");
-
-    if (item.classList.contains('accordion-open')) {
-        accordionContent.removeAttribute('style');
-        item.classList.remove('accordion-open');
-    } else {
-        item.classList.add('accordion-open');
-        accordionContent.style.height = accordionContent.scrollHeight + 'px';
-    }
-}
-
-
-/*=============== SCROLL SECTIONS ACTIVE LINK ===============*/
-const sections = document.querySelectorAll('section[id]');
-// sections.forEach(item => {
-//     console.log(item.offsetHeight, item.offsetTop)
-// })
-
-
-function scrollActive() {
-    const scrollY = window.pageYOffset;
-
-    sections.forEach(current => {
-        const sectionHeight = current.offsetHeight,
-            sectionTop = current.offsetTop - 58,
-            sectionId = current.getAttribute('id');
-
-        if (scrollY > sectionTop && scrollY <= sectionTop + sectionHeight) {
-            document.querySelector('.nav__menu a[href*=' + sectionId + ']').classList.add('active-link');
-        } else {
-            document.querySelector('.nav__menu a[href*=' + sectionId + ']').classList.remove('active-link');
-        }
-    })
-}
-
-
-window.addEventListener('scroll', scrollActive);
-
-
-/*=============== SHOW SCROLL UP ===============*/
-function showScroll() {
-    const scroll = document.querySelector('.scrollup');
-    if (window.pageYOffset > 200) {
-        scroll.classList.add('show-scrollup');
-    } else {
-        scroll.classList.remove('show-scrollup');
-    }
-}
-
-window.addEventListener('scroll', showScroll);
-
-/*=============== DARK LIGHT THEME ===============*/
-const themeButton = document.getElementById('theme-button');
-const darkTheme = 'dark-theme';
-const iconTheme = 'ri-sun-line';
-
-// Previously selected topic (if user selected)
-const selectedTheme = localStorage.getItem('selected-theme');
-const selectedIcon = localStorage.getItem('selected-icon');
-
-
-// We obtain the current theme that the interface has by validating the dark-theme class
-const getCurrentTheme = () => document.body.classList.contains(darkTheme) ? 'dark' : 'light';
-const getCurrentIcon = () => themeButton.classList.contains(iconTheme) ? 'ri-moon-line' : 'ri-sun-line';
-
-
-// We validate if the user previously chose a topic
-if (selectedTheme) {
-    // If the validation is fulfilled, we check to know if we deactivated or activated
-    document.body.classList[selectedTheme === 'dark' ? 'add' : 'remove'](darkTheme);
-    themeButton.classList[selectedIcon === 'ri-moon-line' ? 'add' : 'remove'](iconTheme);
-}
-
-// Activate / deactivate the theme manually with the button
-themeButton.addEventListener('click', () => {
-    // Add or remove dark / icon theme
-    document.body.classList.toggle(darkTheme);
-    themeButton.classList.toggle(iconTheme);
-
-    // Save the current to localStorage
-    localStorage.setItem('selected-theme', getCurrentTheme());
-    localStorage.setItem('selected-icon', getCurrentIcon());
-})
-
-/*=============== SCROLL REVEAL ANIMATION ===============*/
-const sr = ScrollReveal({
-    origin: 'top',
-    distance: '60px',
-    duration: 2500,
-    delay: 400,
-    // reset: true
-})
-
-sr.reveal(`.home__data`);
-sr.reveal(`.home__img`, { delay: 500 });
-sr.reveal(`.home__social`, { delay: 600 });
-sr.reveal(`.about__img, .contact__box`, { origin: 'left' });
-sr.reveal(`.about__data, .contact__form`, { origin: 'right' });
-sr.reveal(`.steps__card, .product__card, .questions__group, .footer`, { interval: 100 });
