--- core/templates/core/home.html
@@ -1,162 +1,0 @@
-
-{%load static%}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<head>
-   <meta charset="UTF-8">
-   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-   <!--=============== FAVICON ===============-->
-   <link rel="shortcut icon" class="ri-command-line" href="/assets/img/favicon.png" type="image/x-icon">
-   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.4/jquery.min.js" integrity="sha512-pumBsjNRGGqkPzKHndZMaAG+bir374sORyzM3uulLV14lN5LyykqNk8eEeUlUkB3U0M4FApyaHraT65ihJhDpQ==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-
-
-   <!--=============== REMIX ICONS ===============-->
-   <link href="https://cdn.jsdelivr.net/npm/remixicon@2.5.0/fonts/remixicon.css" rel="stylesheet">
-
-   <!--=============== CSS ===============-->
-   <link rel="stylesheet" href="{% static 'core/assets/css/style.css' %}">
-
-
-        <!-- +++======++++shit  -->
-
-
-
-    <!-- Bootstrap Notify CSS -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-notify/0.2.0/css/bootstrap-notify.min.css" />
-
-    <!-- Bootstrap Notify JavaScript -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-notify/0.2.0/js/bootstrap-notify.min.js"></script>
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.6-rc.0/css/select2.min.css" rel="stylesheet" />
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.6-rc.0/js/select2.min.js"></script>
-
-
-   <title>Wonder Q || Home Page</title>
-</head>
-</head>
-<body>
-     <header class="header" id="header">
-       <nav class="nav container">
-           <a href="#" class="nav__logo">
-               <i class="ri-question-answer-fill nav__logo-icon"></i>Wonder Q
-           </a>
-
-
-
-           <div class="nav__btns">
-               <!-- Theme change button -->
-               <i class="ri-moon-line change-theme" id="theme-button"></i>
-
-               <div class="nav__toggle" id="nav-toggle">
-                   <i class="ri-menu-line"></i>
-               </div>
-           </div>
-       </nav>
-   </header>
-
-   <main class="main">
-      <!--==================== HOME ====================-->
-      <section class="home" id="home">
-        <div class="home__container container grid">
-            <img src="https://" alt="" class="home__img">
-            <!-- <div class="home__img" >
-                <lottie-player src="https://assets2.lottiefiles.com/packages/lf20_au4zdsr8.json" background="transparent"  speed="1"  style="width: 300px; height: 300px;" loop autoplay ></lottie-player>
-            
-            </div>     -->
-            <div class="home__data">
-                <h1 class="home__title">
-                    Wonder Q <br> 
-                </h1>
-                <p class="home__description">
-                   Ai Tool for self study
-                </p>
-                <p class="home__description">
-                    Generate study questions for your reading material with our Ai powered tool
-                 </p>
-            </div>
-        </div>
-    </section>
-          
-    <section class="contact section container" id="contact">
-        <div class="contact__container grid">
-            <div class="contact__box">
-            <form class="contact__form" name="passageform">
-            {% csrf_token %}
-                <div class="button">
-                    Select File <input type="file" name="pdf_file" accept=".pdf">
-                </div>
-                
-                <div class="button">
-                    <button id="contact" onclick="BtnContact(event)" class="spam-btn button button--flex">
-                        Upload
-                        <i class="ri-arrow-right-up-line button__icon"></i>
-                    </button>
-                </div>
-            </form>
-        </div>
-    </section>
-       <script>
-       function BtnContact(event) {
-        event.preventDefault();
-        var btn = $(event.currentTarget);
-        var originalText = btn.html();
-        btn.prop('disabled', true).html('<i class="fas fa-spinner fa-spin"></i> Processing...');
-        var formData = $(event.currentTarget.form).serialize();
-    
-        console.log(formData)
-        $.ajax({
-            url: "/process/",
-            type: 'POST',
-            data: formData,
-            success: function(response) {
-                if (response.status === 200) {
-            
-                    console.log("success")
-                    console.log(response.questions)
-                    console.log(response.answers)
-
-                    // Save the response to localStorage
-
-                    localStorage.clear();
-
-                    localStorage.setItem('questions', JSON.stringify(response.questions));
-                    localStorage.setItem('answers', JSON.stringify(response.answers));
-
-                    // Redirect the user after a delay
-                    window.location.href = "/details"
-                } else {
-    
-                }
-            },
-            error: function(response) {
-                if (response.status === 500){
-                    console.log(response.error)
-                }
-            
-            },
-            complete: function() {
-                btn.prop('disabled', false).html(originalText);
-            }
-        });
-    }
-    
-
-    </script>
-   </main>
-  
-   <!--=============== SCROLL UP ===============-->
-   <!-- <a href="#" class="scrollup" id="scroll-up">
-       <i class="ri-arrow-up-fill scrollup__icon"></i>
-   </a> -->
-   <!--=============== SCROLL REVEAL ===============-->
-   <!-- <script src="{% static 'core/assets/js/scrollreveal.min.js' %}"></script> -->
-
-   <!--=============== MAIN JS ===============-->
-   <script src="{% static 'core/assets/js/main.js' %}"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.27.2/axios.min.js" integrity="sha512-odNmoc1XJy5x1TMVMdC7EMs3IVdItLPlCeL5vSUPN2llYKMJ2eByTTAIiiuqLg+GdNr9hF6z81p27DArRFKT7A==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-   <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-</body>
-</html>
-
-
