--- hack/settings/production.py
@@ -1,1 +1,0 @@
-from .base import *
