--- hack/wsgi.py
@@ -1,17 +1,0 @@
-"""
-WSGI config for hack project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.2/howto/deployment/wsgi/
-"""
-from dotenv import load_dotenv
-load_dotenv()
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", f"hack.settings.{os.getenv('SETTINGS')}")
-
-application = get_wsgi_application()
