+++ requirements.txt
@@ -19,9 +19,6 @@ defer==1.0.6
 dill==0.3.6
 distro==1.7.0
 distro-info===1.1build1
-dj-database-url==1.3.0
-Django==4.2
-django-jazzmin==2.6.0
 duplicity==0.8.21
 fasteners==0.14.1
 filelock==3.12.0
