--- test.txt
@@ -1,1 +1,0 @@
-chicken wings!!
